+++ tests/testthat/test-fda-table_02.R
@@ -1,16 +1,10 @@
-test_that("make_table_02() works", {
+test_that("fda-table_02() works", {
   skip_if_not_installed("dplyr")
   skip_if_not_installed("gtsummary")
   skip_if_not_installed("pharmaverseadam")
 
-  library(dplyr)
-
-  adsl <- pharmaverseadam::adsl |>
-    filter(TRT01A != "Screen Failure")
-
-  result <- make_table_02(df = adsl)
-
-  ard <- result$ard
+  env <- run_template("fda-table_02")
+  ard <- env$ard
 
   expect_snapshot(as.data.frame(ard$tbl_summary)[1:25, ])
   expect_snapshot(as.data.frame(ard$add_overall)[1:25, ])